iqs62x-h2bin: Sort records based on register address
Sort records in ascending order of register address so that when
the driver writes them in reverse order, the systems setting
register (0xD0) precedes any other fields that may elicit an ATI
cycle when modified.

This ensures that the core clock frequency is reduced before ATI
is triggered later as the driver writes the remaining registers.
If the frequency is reduced while ATI is in progress, the device
may incorrectly report some channels to be in a state of touch.

Also, mask the COMMS_IN_ATI field (bit 3) in the system settings
register such that the device does not request interrupt service
until ATI is complete.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>